změna délka zkratky objektu na 4 znaky, přejmenování duplicitních objektu na Magně (pro přehledenost), výpis mezivozíkového taktu, nějaká ošetření

+++ voziky.h
@@ -65,7 +65,7 @@ public:		// User declarations
 
 		struct TBarva
 		{
-			unsigned int	id_radek;
+			long	id_radek;
 			TColor barva_voziku;
 			struct TBarva *predchozi;
 			struct TBarva *dalsi;
